# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Discodeit API",
    "description": "Sprint Mission",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "전체 User 목록 조회",
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "User 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User 등록",
        "operationId": "create",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "userCreateRequest"
                ],
                "type": "object",
                "properties": {
                  "userCreateRequest": {
                    "$ref": "#/components/schemas/UserCreateRequest"
                  },
                  "profile": {
                    "type": "string",
                    "description": "User 프로필 이미지",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Use가 성공적으로 생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "같은 email 또는 username를 사용하는 User가 이미 존재함",
            "content": {
              "application/json": {
                "example": "User with email {email} already exists"
              }
            }
          }
        }
      }
    },
    "/api/readStatuses": {
      "get": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "User의 Message 읽음 상태 목록 조회",
        "operationId": "findAllByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "조회할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message 읽음 상태 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadStatus"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "Message 읽음 상태 생성",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadStatusCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "message 읽음 상태가 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStatus"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "User 삭제",
        "operationId": "delete",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "삭제할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "User with id {id} not found"
              }
            }
          },
          "204": {
            "description": "User가 성공적으로 삭제됨"
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "User 정보 수정",
        "operationId": "update",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "수정할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "userUpdateRequest"
                ],
                "type": "object",
                "properties": {
                  "userUpdateRequest": {
                    "$ref": "#/components/schemas/UserUpdateRequest"
                  },
                  "profile": {
                    "type": "string",
                    "description": "수정할 User 프로필 이미지",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "같은 email 또는 username를 사용하는 User가 이미 존재함",
            "content": {
              "*/*": {
                "example": "user with email {newEmail} already exists"
              }
            }
          },
          "200": {
            "description": "User 정보가 성공적으로 수정됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "User with id {userId} not found"
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/userStatus": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "사용자 상태 변경",
        "operationId": "updateUserStatusByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "상태를 변경할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "해당 User의 UserStatus를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "UserStatus with userId {userId} not found"
              }
            }
          },
          "200": {
            "description": "User온라인 상태가 성공적으로 업데이트됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/readStatuses/{readStatusId}": {
      "patch": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "Message 읽음 상태 수정",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "readStatusId",
            "in": "path",
            "description": "수정할 읽음 상태",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Message 읽음 상태를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "ReadStatus with id {readStatusId} not found"
              }
            }
          },
          "200": {
            "description": "Message 읽음 상태가 성공적으로 수정됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStatus"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserCreateRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "description": "User 생성 정보"
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "profileId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ReadStatusCreateRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReadStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "newUsername": {
            "type": "string"
          },
          "newEmail": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "description": "수정할 User 정보"
      },
      "UserStatusUpdateRequest": {
        "type": "object",
        "properties": {
          "newLastActiveAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "lastConnectedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "online": {
            "type": "boolean"
          }
        }
      },
      "ReadStatusUpdateRequest": {
        "type": "object",
        "properties": {
          "newLastReadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "profileId": {
            "type": "string",
            "format": "uuid"
          },
          "online": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}